% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_if_null}
\alias{check_if_null}
\title{Checks only 1 option has been set}
\usage{
check_if_null(
  ...,
  stopIfNoNull = TRUE,
  noNullMessage = "No conditions selected!"
)
}
\arguments{
\item{...}{The arguments to check for NULL}

\item{stopIfNoNull}{LOGICAL: If TRUE, an error is thrown; if FALSE, a
warning is thrown.}

\item{noNullMessage}{"string" The warning or error message.}
}
\description{
If a function has multiple arguments that default to NULL,
this function checks that only one of them has been set at once.
}
