% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_by_cell.R
\name{correlate_by_cell}
\alias{correlate_by_cell}
\title{Calculates the per-pixel correlation between 2 SpatRasters}
\usage{
correlate_by_cell(xy, detrend, method = "pearson")
}
\arguments{
\item{xy}{SpatRaster: The dataset, which should be \code{c(x, y)}.}

\item{detrend}{BINARY: Should the time series be detrended before the
correlation?}

\item{method}{"string": Which correlation coefficient to use? One of
"pearson", "kendall", or "spearman". See \code{stats::cor.test()}.}
}
\description{
Say we have a temperature SpatRaster with 20 layers, and a
precipitation SpatRaster with 20 layers. We want to create a map that
shows the correlation between temperature and precipitation across the 20
years for \emph{each} pixel. Use this function and \code{terra::app()}. This
function largely follows the code from Robert Hijmans here:
https://stackoverflow.com/questions/16698048/r-calculating-pearson-correlation-coefficient-in-each-cell-along-time-line

It is necessary to first combine the two datasets, using \code{c()}, and then
run the \code{terra::app()} function. See examples.
}
\examples{
  \dontrun{
    # We want to correlate x with y
    x <- terra::rast("pretend1.nc")   # n layers
    y <- terra::rast("pretend2.nc")   # also n layers

    # Combine
    xy <- c(x, y)

    # Correlate using terra::app
    koral <- terra::app(x = xy,
                        detrend = TRUE,
                        method  = "pearson",
                        fun = correlate_by_cell)

    names(koral) <- c("estimate", "pValue")
  }

}
