% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_unmatched.R
\name{exclude_unmatched_months}
\alias{exclude_unmatched_months}
\title{Subset a SpatRaster to exclude layers when the month isn't found in all years}
\usage{
exclude_unmatched_months(x, australSplit = NULL, printClarity = FALSE)
}
\arguments{
\item{x}{The SpatRaster data to subset.}

\item{australSplit}{numeric: If not NULL, create an additional column for the
austral summer. For example, December 1991 and January 1992 should often
be considered as part of summer 1991/1992 rather than in their respective
years. The default value is 3, which means that all months \emph{AFTER} March
are considered as part of the following summer (i.e. April 1991 -- March
1992 are all in 1992). Swap this value according: setting it as 4 means
May 1991 -- April 1992 are all 1992.}

\item{printClarity}{LOGICAL: If TRUE, print out an explanation when the
function runs to clarify which data is being excluded and which included.}
}
\description{
The 4 \code{exclude_x} functions are easiest to explain with
examples. Be very careful and deliberate with these functions to know
exactly what is being excluded.

Take a dataset containing monthly data for Jan-Apr in 1980 and 1981,
Jan--Mar in 1982, and Jan--Feb in 1983. If mean values were calculated on
a monthly basis, the January and February values would be calculated from
4 layers each (i.e. 1980, 1981, 1982, and 1983), but the March value
would be from 3 layers (1980, 1981, 1982), and the April data from just
two layers (1980, 1981). The \code{exclude_x} functions can help in a few ways
to wrangle the data before running the means.
\subsection{Example 1}{

Using \code{exclude_incomplete_years}, the 1982 and 1983 data is
excluded. These two years do not contain all of the months found in the
data (i.e. Jan, Feb, Mar, Apr): neither 1982 or 1983 have April data, and
1983 does not have March data either. Only data layers in years in which
all of the months in the dataset (i.e. Jan--April) occur are retained:
here, that would be 1980 and 1981, which both have data for Jan--Apr.

The term "incomplete_years" is therefore relative; it is defined by the
data fed in. A year in this function is defined with calendar dates (i.e.
January through December); for the equivalent with austral years /
summers, use \code{exclude_incomplete_summers}.

This example is based on a monthly resolution (i.e. January, not e.g.
January 16th). A monthly resolution in this function is particularly
useful for monthly resolution data - all data layers in terra still
require a full date, but the day part is essentially an arbitrary number
(e.g. it could refer to the midpoint or end of the month, which depend on
the length of the month). Set the \code{daily} argument as TRUE to use this
function with daily resolution data (i.e. when the day part of the date
is important). If daily resolution, 29th February is ignored if missing,
and years/summers without it are still included.
}

\subsection{Example 2}{

Taking the same initial dataset, and using
\code{exclude_unmatched_months}, the March and April data is excluded. These
two months do not have data in all of the years found in the data (i.e.
1980, 1981, 1982, 1983): April data doesn't exist in either 1982 or 1983,
and March data doesn't exist in 1982. Only data layers in months found in
all years of the dataset (i.e. 1980:1983) are retained: here, that would
be January and February, both of which have data in 1980:1983.

For the equivalent function, but using both the day and month part of the
date, use \code{exclude_unmatched_days}. Both of the \code{exclude_unmatched}
functions can be run for a calendar year ('australSplit = NULL') or for
an austral year / summer.
}
}
