% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_dates.R
\name{handle_dates}
\alias{handle_dates}
\title{Reformat date information}
\usage{
handle_dates(x, out = "YYYY-MM-DD", outSep = "-")
}
\arguments{
\item{x}{A vector containing strings in which the month is clear. For a
list, use \code{handled <- lapply(x, handle_dates)}.}

\item{out}{How should the date be returned? Choose from any of the
following (using the example of 20th June 1991).

\if{html}{\out{<div class="sourceCode">}}\preformatted{    "YYYY-MM-DD"     returns     "1991-06-20"
    "YMD"            returns     "1991-06-20"
    "DD-MM-YYYY"     returns     "20-06-1991"
    "DMY"            returns     "20-06-1991"
    "YYYY-Jan-DD"    returns     "1991-Jun-20"
    "YBD"            returns     "1991-Jun-20"
    "DD-Jan-YYYY"    returns     "20-Jun-1991"
    "DBY"            returns     "20-Jun-1991"
}\if{html}{\out{</div>}}

Input is case-insensitive. The divider can be set using the 'outSep'
argument.}

\item{outSep}{Which symbol should be placed between the date and month?
Default is "-", but will accept any reasonable string.}
}
\description{
Sometimes dates are represented as strings, potentially with
the month in the format "Dec" for example. This function handles a few of
these and allows them to be reformatted. It is conceptually similar to
the \code{as.Date} and \code{strptime} functions, but doesn't require the data to
actually be a date.
}
