% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_monthDays.R
\name{handle_monthDays}
\alias{handle_monthDays}
\title{Reformat "month-day" information}
\usage{
handle_monthDays(x)
}
\arguments{
\item{x}{A vector containing strings in which the month is clear. For a
list, use \code{handled <- lapply(x, handle_monthDays)}.}
}
\description{
This function reformats month-day combinations to match
"Jun-04".

This function should work as long as the month is unambiguous in a string
with a reasonable separator (see below). It can handle these inputs:

\if{html}{\out{<div class="sourceCode">}}\preformatted{   01-Jan  | 1-Jan | Jan-1 | Jan-01
   01/Jan  | 1/Jan | Jan/1 | Jan/01
   01 Jan  | 1 Jan | Jan 1 | Jan 01
   01_Jan  | 1_Jan | Jan_1 | Jan_01

The Jan can also be "jan", "January", or "jan".
}\if{html}{\out{</div>}}

\strong{note} This function cannot handle unambiguous combinations (e.g.
02/01). It also doesn't (currently) strip out any year information in
the date.
}
