% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_SpatRaster.R
\name{return_SpatRaster}
\alias{return_SpatRaster}
\title{Handle whether x is a filename or SpatRaster}
\usage{
return_SpatRaster(x)
}
\arguments{
\item{x}{The input to check. If a string, it is fed into \code{\link[terra:rast]{terra::rast()}} on
the assumption it is a file path. If a SpatRaster, returned as is.}
}
\description{
Handle whether x is a filename or SpatRaster
}
