% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_by.R
\name{subset_by}
\alias{subset_by}
\title{Subset a SpatRaster based on the layers' dates}
\usage{
subset_by(
  x,
  type,
  exact = NULL,
  before = NULL,
  after = NULL,
  except = NULL,
  australSplit = 3
)
}
\arguments{
\item{x}{SpatRaster: The data to subset. It can be either a string, in
which case it is interpreted as a filepath and read in using
\code{\link[terra:rast]{terra::rast()}}; or an existing SpatRaster.}

\item{type}{Which time/date type should the data be subset based on?
Options are (example using a date of 2019-11-29 and time of 12:30):

\if{html}{\out{<div class="sourceCode">}}\preformatted{- "date"          2019-11-29
- "year"          2019
- "month"         11
- "day"           29
- "monthDay"      Nov-29
- "summer"        2020
- "hour"          12
- "minute"        30
}\if{html}{\out{</div>}}}

\item{exact}{vector: Which data should be included in the returned
SpatRaster? Only data that occurs on these exact dates / times is
returned. Cannot be used in conjunction with the 'before', 'after' or
'except' arguments.}

\item{before}{Only data that occurs before this value is returned. Cannot
be used in conjunction with the exact values, 'after' or 'except'
arguments.}

\item{after}{Only data that occurs after this value is returned. Cannot be
used in conjunction with the exact values, 'before' or 'except'
arguments.}

\item{except}{vector: Which data should be excluded from the SpatRaster?
Cannot be used in conjunction with the exact values, 'after' or 'before'
arguments.}

\item{australSplit}{numeric: If not FALSE, create an additional column for
the austral summer / year. For example, December 1991 and January 1992
often need to be considered as part of summer 1991/1992 in the southern
hemisphere rather than in their respective years.

The default value is 3, which means that all months \emph{AFTER} March (i.e.
month 3) are considered as part of the following austral summer / year
(i.e. April 1991 -- March 1992 are all in austral summer / year 1992).

Swap this value accordingly: e.g. setting it as 4 means May 1991 -- April
1992 are all austral summer / year 1992.}
}
\description{
Easily select layers of a SpatRaster depending on their date
(and time, if present). This was designed as an internal function that
the other \code{subset_by_x} functions are based on but is also available on
its own.
}
