% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_by_day.R
\name{subset_by_day}
\alias{subset_by_day}
\title{Subset a SpatRaster based on the layers' day}
\usage{
subset_by_day(x, days = NULL, before = NULL, after = NULL, except = NULL)
}
\arguments{
\item{x}{SpatRaster: The data to subset. It can be either a string, in
which case it is interpreted as a filepath and read in using
\code{\link[terra:rast]{terra::rast()}}; or an existing SpatRaster.}

\item{days}{Which day/s to return? Use this argument for exact matches
(e.g. c(1, 8, 15:18). Cannot be used in conjunction with the 'before',
'after' or 'except' arguments.}

\item{before}{Only data that occurs before this value is returned. Cannot
be used in conjunction with the exact values, 'after' or 'except'
arguments.}

\item{after}{Only data that occurs after this value is returned. Cannot be
used in conjunction with the exact values, 'before' or 'except'
arguments.}

\item{except}{vector: Which data should be excluded from the SpatRaster?
Cannot be used in conjunction with the exact values, 'after' or 'before'
arguments.}
}
\description{
Easily select only the layers of a SpatRaster depending on
only the layers' day. For example, only data on the 1st of a month (1),
or only data between the 10th and 15th of a month (10:15).

This function is distinct to \code{subset_by_date()} because it subsets
using \emph{only} the day part of the date: there is no accounting for the
month or year. This approach can be useful to select the first day of
each month for example.
}
