% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_by_hour.R
\name{subset_by_hour}
\alias{subset_by_hour}
\title{Subset a SpatRaster based on the layers' hour}
\usage{
subset_by_hour(x, hours = NULL, before = NULL, after = NULL, except = NULL)
}
\arguments{
\item{x}{SpatRaster: The data to subset. It can be either a string, in
which case it is interpreted as a filepath and read in using
\code{\link[terra:rast]{terra::rast()}}; or an existing SpatRaster.}

\item{hours}{Which hour/s to return? Use this argument for exact matches
(e.g. c(12:17, 21-22)), otherwise leave this as NULL (the default) and
use one of the other arguments. Must be in the 24-hour format.}

\item{before}{numeric: The maximum value that the 'type' value can have.
Only one of the criteria arguments can be used at once.}

\item{after}{numeric: The minimum value that the 'type' value can have.
Only one of the criteria arguments can be used at once.}

\item{except}{vector: Values that the 'type' value cannot match.}
}
\description{
Easily select only the layers of a SpatRaster depending on
only the layers' hour. For example, only data at 12:00, or only data
after 17:00. Works only on the hour (ignores minutes!) and if time
information is available in the 'x' raster.
}
