% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_by_minute.R
\name{subset_by_minute}
\alias{subset_by_minute}
\title{Subset a SpatRaster based on the layers' minute}
\usage{
subset_by_minute(x, minutes = NULL, before = NULL, after = NULL, except = NULL)
}
\arguments{
\item{x}{SpatRaster: The data to subset. It can be either a string, in
which case it is interpreted as a filepath and read in using
\code{\link[terra:rast]{terra::rast()}}; or an existing SpatRaster.}

\item{minutes}{Which minutes/s to return? Use this argument for exact
matches (e.g. c(0:15, 30, 45). Cannot be used in conjunction with
the 'before', 'after' or 'except' arguments.}

\item{before}{Only data that occurs before this value is returned. Cannot
be used in conjunction with the exact values, 'after' or 'except'
arguments.}

\item{after}{Only data that occurs after this value is returned. Cannot be
used in conjunction with the exact values, 'before' or 'except'
arguments.}

\item{except}{vector: Which data should be excluded from the SpatRaster?
Cannot be used in conjunction with the exact values, 'after' or 'before'
arguments.}
}
\description{
Easily select only the layers of a SpatRaster depending on
only the layers' minute. For example, only data at quarter past the hour
(15) or only data in the last half of an hour (>30). Works only on the
minute (ignores hours!) and if time information is available in the 'x'
raster.
}
