% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_by_monthDay.R
\name{subset_by_monthDay}
\alias{subset_by_monthDay}
\title{Subset a SpatRaster based on the layers' month-day}
\usage{
subset_by_monthDay(
  x,
  monthDays = NULL,
  periods = NULL,
  excludeIncomplete = FALSE,
  before = NULL,
  after = NULL,
  except = NULL
)
}
\arguments{
\item{x}{SpatRaster: The data to subset. It can be either a string, in
which case it is interpreted as a filepath and read in using
\code{\link[terra:rast]{terra::rast()}}; or an existing SpatRaster.}

\item{monthDays}{Which monthDay/s to return?  Used for specific monthDays.
Cannot be used in conjunction with the 'periods' argument.

Input must be a vector of specific monthDay/s, best formatted as "Jan-14"
or "14-Jan". The month must be explicit.

As an example, using a vector of \verb{c("Jan-01", "7 Feb", "15/Mar)} will
find any layers that occur on the 1st of January or the 7th of February
or the 15th March in the input SpatRaster, regardless of the layers'
year.

\strong{Note:} Something like "01/02" will fail. It is ambiguous. This
argument is fed directly into \code{\link[=handle_monthDays]{handle_monthDays()}} to identify
and reformat the month-day, so see that function for details on accepted
inputs.}

\item{periods}{Which monthDay/s to return? Used for extended periods rather
than individual monthDays. Cannot be used in conjunction with the
'monthDays' argument.

Input dates must be a vector, with the first value indicating when the
period begins and the second when the period ends.

For multiple periods, use a list of such vectors, for example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(c("Jan-01", "Jan-01"),   # 1-3  Jan for every year in the SpatRaster
     c("Mar-08", "Mar-11"))   # 8-11 Mar for every year in the SpatRaster
}\if{html}{\out{</div>}}

The above will return data from the 1st, 2nd and 3rd of January, and the
8th, 9th, 10th and 11th of March, regardless of the year.}

\item{before}{Only data that occurs before this value is returned. Cannot
be used in conjunction with the exact values, 'after' or 'except'
arguments.}

\item{after}{Only data that occurs after this value is returned. Cannot be
used in conjunction with the exact values, 'before' or 'except'
arguments.}

\item{except}{Which monthDay/s should be excluded from the SpatRaster? If a
vector, all dates are treated individually; for extended periods, use a
list of vectors as outlined in the 'periods' argument.}
}
\description{
Easily select only layers of a SpatRaster that are on certain
month-days (i.e. dates which ignore the year).

As an example, return only the layers that fall on the 1st, 2nd or 3rd of
December for every year in the dataset.
}
