% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_by_year.R
\name{subset_by_year}
\alias{subset_by_year}
\title{Subset a SpatRaster based on the layers' year}
\usage{
subset_by_year(x, years = NULL, before = NULL, after = NULL, except = NULL)
}
\arguments{
\item{x}{SpatRaster: The data to subset. It can be either a string, in
which case it is interpreted as a filepath and read in using
\code{\link[terra:rast]{terra::rast()}}; or an existing SpatRaster.}

\item{years}{Which year/s to return? Use this argument for exact matches
(e.g. c(1983:1986, 1991:1992). Cannot be used in conjunction with
the 'before', 'after' or 'except' arguments.}

\item{before}{Only data that occurs before this value is returned. Cannot
be used in conjunction with the exact values, 'after' or 'except'
arguments.}

\item{after}{Only data that occurs after this value is returned. Cannot be
used in conjunction with the exact values, 'before' or 'except'
arguments.}

\item{except}{vector: Which data should be excluded from the SpatRaster?
Cannot be used in conjunction with the exact values, 'after' or 'before'
arguments.}
}
\description{
Easily select only the layers of a SpatRaster depending on
only the layers' year. For example, only data in 1991, or only data after
1999.
}
