% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_terra_dates.R
\name{get_date_info}
\alias{get_date_info}
\title{Return a data frame of SpatRaster layer dates for subsetting}
\usage{
get_date_info(x, australSplit = 3)
}
\arguments{
\item{x}{SpatRaster: The dataset.}

\item{australSplit}{numeric: If not FALSE, create an additional column for
the austral summer / year. For example, December 1991 and January 1992
often need to be considered as part of summer 1991/1992 in the southern
hemisphere rather than in their respective years. The default value is 3,
which means that all months \emph{AFTER} March (i.e. month 3) are considered
as part of the following austral summer / year (i.e. April 1991 -- March
1992 are all in austral summer / year 1992). Swap this value accordingly:
e.g. setting it as 4 means May 1991 -- April 1992 are all austral summer
/ year 1992.}
}
\description{
Often it is necessary to select only a subset of layers in a
dataset based on their date or time. For example, only layers in
December, only in 2017, or only in the afternoon. The date (and possibly
time) information is stored in the SpatRaster; this function creates a
simple data frame to hold the date information, thus allowing the correct
layers to be identified and filtered. This only works with SpatRasters
from the \code{terra} package, and they must have the date & time information
necessary. Any missing data is assigned a value of NA. If the data has
no time information, the time columns are excluded.
}
