% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_by_date.R
\name{subset_by_date}
\alias{subset_by_date}
\title{Subset a SpatRaster based on the layers' date}
\usage{
subset_by_date(x, dates = NULL, dateList = NULL)
}
\arguments{
\item{x}{SpatRaster: The data to subset. Can be either a string, in which
case it is interpreted as a filePath and read in, or an existing
SpatRaster.}

\item{dates}{A vector of specific dates to return. Must be in the format
YYYY-MM-DD.}

\item{dateList}{A list of vectors, with the first value in each vector
indicating when the period begins, and the second when the period ends.}
}
\description{
Easily select only layers of a SpatRaster that are on specific
  dates. This function is distinct to `subset_by_day()` as it considers the
  full date, not just the day. It is usually quicker to chain together the
  "subset_by_x" functions unless an unusual combination of dates are
  required. See the examples.

  The dates can be input either as a vector of specific dates using the
  "dates" argument; or as a list of vectors containing the start and end
  dates for longer extended periods. See examples.
}
\examples{
\dontrun{
  x <- terra::rast("fileName.nc")

  # For specific dates, use the "dates" argument:
  xSubset1 <- subset_by_date(x, dates = c("1991-12-01", "1994-12-19",
                                          "2003-04-17", "2012-09-22"))

  # For extended periods, use the "dateList" argument:
  xSubset2 <- subset_by_date(x,
                             dateList = list(c("1991-12-01", "1991-12-05"),
                                             c("1993-11-15", "1993-12-15")))

  # However, it can be easier to chain other "subset_by_x" functions:
  xSubset3 <- subset_by_year(x, 1991) |> subset_by_month(12)

  # instead of using subset_by_date:
  xSubset4 <- subset_by_date(x,
                             dateList = list(c("1991-12-01", "1991-12-31")
}
}
