% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_by_day.R
\name{subset_by_day}
\alias{subset_by_day}
\title{Subset a SpatRaster based on the layers' day}
\usage{
subset_by_day(x, days)
}
\arguments{
\item{x}{SpatRaster: The data to subset. Can be either a string, in which
case it is interpreted as a filePath and read in, or an existing
SpatRaster.}

\item{days}{vector: Which day/s to return?}
}
\description{
Easily select only layers of a SpatRaster that are on a
  certain day/s. For example, only want layers that are on the 1st of a
  month. This function is distinct to `subset_by_date()` as it subsets
  using *only* the day part of the date: there is no accounting for the
  month or year. This approach can be useful to select the first day of
  each month for example.
}
