% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_by_day.R
\name{subset_by_day}
\alias{subset_by_day}
\title{Subset a SpatRaster based on the layers' day}
\usage{
subset_by_day(x, days = NULL, before = NULL, after = NULL, except = NULL)
}
\arguments{
\item{x}{SpatRaster: The data to subset. It can be either a string, in
which case it is interpreted as a filepath and read in using
\code{\link[terra:rast]{terra::rast()}}; or an existing SpatRaster.}

\item{days}{Which day/s to return? Use this argument for exact matches
(e.g. c(1, 8, 15)), otherwise leave this as NULL (the default) and use
one of the other arguments.}

\item{before}{numeric: The maximum value that the 'type' value can have.
Only one of the criteria arguments can be used at once.}

\item{after}{numeric: The minimum value that the 'type' value can have.
Only one of the criteria arguments can be used at once.}

\item{except}{vector: Values that the 'type' value cannot match.}
}
\description{
Easily select only the layers of a SpatRaster depending on
only the layers' day. For example, only data on the 1st of a month (1),
or only data between the 10th and 15th of a month (10:15).

This function is distinct to \code{subset_by_date()} because it subsets
using \emph{only} the day part of the date: there is no accounting for the
month or year. This approach can be useful to select the first day of
each month for example.
}
