% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_by.R
\name{subset_by_minute}
\alias{subset_by_minute}
\title{Subset a SpatRaster based on the layers' minute}
\usage{
subset_by_minute(
  x,
  minutes = NULL,
  before = NULL,
  after = NULL,
  between = NULL,
  except = NULL
)
}
\arguments{
\item{x}{SpatRaster: The data to subset. It can be either a string, in
which case it is interpreted as a filepath and read in using
\code{\link[terra:rast]{terra::rast()}}; or an existing SpatRaster.}

\item{before}{numeric: The maximum value that the 'type' value can have.
Only one of these 5 arguments can be used at once.}

\item{after}{numeric: The minimum value that the 'type' value can have.
Only one of these 5 arguments can be used at once.}

\item{between}{vector: The minimum and maximum values that the 'type' value
can have; inclusive.}

\item{except}{vector: Values that the 'type' value cannot match.}

\item{hours}{Which minute/s to return? Use this argument for exact matches
(e.g. c(0:15, 30:40)), otherwise leave this as NULL (the default) and use
one of the other arguments.}
}
\description{
Easily select only the layers of a SpatRaster depending on
only the layers' minute. For example, only data at quarter past the hour
(15) or only data in the last half of an hour (>30). Works only on the
minute (ignores hours!) and if time information is available in the 'x'
raster.
}
