% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_by_month.R
\name{subset_by_month}
\alias{subset_by_month}
\title{Subset a SpatRaster based on the layers' month}
\usage{
subset_by_month(x, months, removeIncomplete = FALSE, dailyResolution = FALSE)
}
\arguments{
\item{x}{SpatRaster: The data to subset. Can be either a string, in which
case it is interpreted as a filePath and read in, or an existing
SpatRaster.}

\item{months}{vector: Which month/s to return? Input as either the month
number (12), the full month name ("December" or "december"), or the
abbreviated month name ("Dec" or "dec"). Multiple months can be input at
once (e.g. c(12, 1, 2)), but do not try to mix strings and numbers in the
vector.}

\item{removeIncomplete}{If the value is "years", the data is run through
\code{remove_incomplete_years()}, and only months in years with all requested
months in are returned. If the value is numeric (between 1 and 12), the
data is fed into the \code{remove_incomplete_austral_summers()} and the value
is used as the "australSplit" argument to return only months in austral
summers that contain all requested months. If any other value, all layers
matching the months argument are returned, regardless of the summer or
year. See the examples.}

\item{dailyResolution}{BINARY: If using the "removeIncomplete" argument, it
is necessary to define whether the data is at a daily or monthly
resolution. See the \code{remove_incomplete_austral_summers()} and
\code{remove_incomplete_years()} functions for more information.}
}
\description{
Easily select only layers of a SpatRaster that are in a
certain month/s (e.g. only layers that are in December).
}
