% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_by_monthDay.R
\name{subset_by_monthDay}
\alias{subset_by_monthDay}
\title{Subset a SpatRaster based on the layers' month-day}
\usage{
subset_by_monthDay(
  x,
  monthDays = NULL,
  monthDayList = NULL,
  excludeIncomplete = FALSE
)
}
\arguments{
\item{x}{SpatRaster: The data to subset. Can be either a string, in which
case it is interpreted as a filePath and read in, or an existing
SpatRaster.}

\item{monthDays}{Used for specific dates. A vector of monthDays, best
formatted as "Jan-14" or "14-Jan". Fed directly into \code{\link[=handle_monthDays]{handle_monthDays()}}
to identify and reformat the month-day for matching against the
SpatRaster dates.

As an example, using a vector of \verb{c("Jan-01", "7 Feb", "15/Mar)} will
find any layers that occur on the 1st of January or the 7th of February
or the 15th March in the 'x' dataset, regardless of the layers' year.

\strong{Note:} Something like "01/02" will fail. It is ambiguous.}

\item{monthDayList}{Used for extended periods of dates. Input must be a
list of vectors, with the first value in each vector indicating when the
period begins, and the second when the period ends.

For example, \code{list(c("Jan-02", "Jan-04"), c("Feb-07", "Feb-09"))} will
return all layers that occur on the 2nd, 3rd or 4th January, or on the
7th, 8th or 9th of February in the dataset, regardless of the layers'
year.}

\item{excludeIncomplete}{Be careful using this argument! It can
dramatically affect the output of this function. Please read the
Explanation in \code{\link[=exclude_incomplete_years]{exclude_incomplete_years()}} first.

If TRUE, the data is run through \code{exclude_incomplete_years()}, and only
months in years with all requested months in are returned.

If numeric (between 1 and 12), the data is fed into
\code{exclude_incomplete_summers()} and the value is used as the
'australSplit' argument to return only months in austral summers that
contain all requested months.

If any other value (including the default FALSE), the
\code{exclude_incomplete_x} functions are skipped, and all layers matching the
'months' argument are returned, regardless of the summer or year in which
they occur. See the examples.}
}
\description{
Easily select only layers of a SpatRaster that are on certain
month-days (i.e. dates ignoring the year).

As an example, return only the layers that fall on the 1st, 2nd or 3rd of
December for every year in the dataset.
}
