% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_by_monthDay.R
\name{subset_by_monthDay}
\alias{subset_by_monthDay}
\title{Subset a SpatRaster based on the layers' month-day}
\usage{
subset_by_monthDay(
  x,
  monthDays = NULL,
  periods = NULL,
  excludeIncomplete = FALSE,
  before = NULL,
  after = NULL,
  except = NULL
)
}
\arguments{
\item{x}{SpatRaster: The data to subset. It can be either a string, in
which case it is interpreted as a filepath and read in using
\code{\link[terra:rast]{terra::rast()}}; or an existing SpatRaster.}

\item{monthDays}{Used for specific dates. A vector of monthDays, best
formatted as "Jan-14" or "14-Jan". Fed directly into \code{\link[=handle_monthDays]{handle_monthDays()}}
to identify and reformat the month-day for matching against the 'x'
SpatRaster dates.

As an example, using a vector of \verb{c("Jan-01", "7 Feb", "15/Mar)} will
find any layers that occur on the 1st of January or the 7th of February
or the 15th March in the 'x' dataset, regardless of the layers' year.

\strong{Note:} Something like "01/02" will fail. It is ambiguous.}

\item{periods}{Used for extended periods rather than individual dates.
Input dates must be a vector, with the first value indicating when the
period begins and the second when the period ends. For multiple periods,
use a list of such vectors, for example:
```
list(c("Jan-01", "Jan-01"),     # 1-3  Jan of all years in x
c("Mar-08", "Mar-11"))     # 8-11 Mar of all years in x
````
The above will return the layers for 1st, 2nd and 3rd of January, and the
8th, 9th, 10th and 11th of March, regardless of the year.}

\item{before}{numeric: The maximum value that the 'type' value can have.
Only one of the criteria arguments can be used at once.}

\item{after}{numeric: The minimum value that the 'type' value can have.
Only one of the criteria arguments can be used at once.}

\item{except}{Dates that should be removed from the 'x' SpatRaster. If a
vector, all dates are treated individually; for extended periods, use a
list of vectors as outlined in the 'periods' argument.}
}
\description{
Easily select only layers of a SpatRaster that are on certain
month-days (i.e. dates which ignore the year).

As an example, return only the layers that fall on the 1st, 2nd or 3rd of
December for every year in the dataset.
}
