% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_by_year.R
\name{subset_by_year}
\alias{subset_by_year}
\title{Subset a SpatRaster based on the layers' year}
\usage{
subset_by_year(x, years)
}
\arguments{
\item{x}{SpatRaster: The data to subset. Can be either a string, in which
case it is interpreted as a filePath and read in; or an existing
SpatRaster.}

\item{years}{vector: Which year/s to return?}
}
\description{
Easily select only layers of a SpatRaster that are in a
  certain year/s. For example, only want layers that are in 1991.
}
\examples{
\dontrun{
  x <- terra::rast("fileName.nc")

  # Only 1991
  x91 <- terrapin::subset_by_year(x, 1991)

  # From 1991 to 1994
  x9194 <- terrapin::subset_by_year(x, 1991:1994)
}

}
